Move examples out of the main library
