+++ tests/test_canvas.py
@@ -1,7 +1,7 @@
 import pytest
 
 from gaphas.canvas import Canvas, ConnectionError
-from gaphas.examples import Box
+from gaphas.item import Element as Box
 from gaphas.item import Line
 from gaphas.matrix import Matrix
